Only run URL check if not on CI
